\name{Rnvim.source}
\alias{Rnvim.source}
\title{Wrapper to base::source}
\description{
  Call base::source with the arguments \code{print.eval=TRUE} and
  \code{spaced=FALSE}.
}
\usage{
  Rnvim.source(..., print.eval = TRUE, spaced = FALSE)
}
\arguments{
  \item{print.eval}{See base::source.}
  \item{spaced}{See base::source.}
  \item{...}{Further arguments passed to base::source.}
}
